+++ src/components/EmptyState/index.tsx
@@ -8,7 +8,14 @@
  */
 
 import * as React from 'react';
-import { Button, ButtonProps, Grid, Typography } from '@mui/material';
+import {
+  Button,
+  ButtonProps,
+  Grid,
+  IconProps,
+  Typography,
+  TypographyProps,
+} from '@mui/material';
 import clsx from 'clsx';
 import { isString } from 'lodash';
 
@@ -17,17 +24,25 @@ import useStyles from './styles';
 export interface EmptyStateProps {
   buttonProps?: ButtonProps;
   description?: any;
+  descriptionTypographyProps?: TypographyProps;
   Icon?: React.FC<any> | React.ComponentClass<any>;
+  iconProps?: IconProps;
   style?: React.CSSProperties;
   title?: string | React.ReactElement<unknown>;
+  titleTypographyProps?: TypographyProps & {
+    component?: string;
+  };
 }
 
 export function EmptyState({
   buttonProps,
   description,
+  descriptionTypographyProps,
   Icon,
+  iconProps,
   style,
   title,
+  titleTypographyProps,
 }: EmptyStateProps): React.ReactElement<EmptyStateProps> {
   const classes = useStyles();
 
@@ -40,6 +55,7 @@ export function EmptyState({
           }}
           color="disabled"
           titleAccess={isString(title) ? title : undefined}
+          {...iconProps}
         />
       )}
 
@@ -58,7 +74,15 @@ export function EmptyState({
             }}
             item
           >
-            <Typography align="center" variant="h6">
+            <Typography
+              align="center"
+              classes={{
+                root: classes.titleRoot,
+              }}
+              component="h6"
+              variant="body1"
+              {...titleTypographyProps}
+            >
               {title}
             </Typography>
           </Grid>
@@ -74,7 +98,9 @@ export function EmptyState({
               !title && classes.descriptionRootWithoutTitle,
             ),
           }}
+          color="textSecondary"
           variant="body1"
+          {...descriptionTypographyProps}
         >
           {description}
         </Typography>